// Web/RMA-Web-Shell/projects/clientcare/src/app/broker-manager/models/representative-bank-account.ts
export class RepresentativeBankAccount {
    id: number;
    representativeId: number;
    effectiveDate: Date;
    accountNumber: string;
    bankBranchId: number;
    bankAccountType: number;
    accountHolderName: string;
    branchCode: string;
    approvalRequestedFor: string;
    approvalRequestId: number;
    isApproved: boolean;
    reason: string;
    statusText: string;
    isDeleted: boolean;
    createdBy: string;
    createdDate: Date;
    modifiedBy: string;
    modifiedDate: Date;
  }